requirements.txt
pycryptodome~=3.21.0
pysftp~=0.2.9